requirements.txt
# ViT + XGBoost 육류 품질 예측 모델 의존성

# 딥러닝 및 트랜스포머
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.15.0

# 머신러닝
xgboost>=1.5.0
scikit-learn>=1.0.2

# 데이터 처리
pandas>=1.3.0
numpy>=1.21.0

# 이미지 처리
Pillow>=8.3.0
opencv-python>=4.5.0

# 설정 및 유틸리티
PyYAML>=6.0
tqdm>=4.62.0

# 시각화 (선택사항)
matplotlib>=3.5.0
seaborn>=0.11.0

# 통계 및 메트릭
scipy>=1.7.0

# 추가 유틸리티
openpyxl>=3.0.9  # Excel 파일 처리
sqlite3  # 기본 내장됨